refactor: make facility card more flexible
